+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- prints a message when Mokoki has not been able to get the `StackTraceElement`
+
 ### Changed
 
 ### Deprecated
@@ -12,17 +14,23 @@
 
 ### Fixed
 
+- iOS was throwing an exception instead of `null`
+
 ### Updated
+
 - Gradle Wrapper to 7.0
 - Massive Catalogs to 0.1.0-alpha.22
 
 ## [0.1.0-alpha.2] - 2021-03-26
 
 ### Added
+
 - Add iOS target
 
 ### Changed
+
 - Rename `mokoki` subproject to `mokoki-core`
 
 ## [0.1.0-alpha.1] - 2021-03-22
+
 - Initial release